apps/ai-services/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pillow==11.0.0
numpy==1.26.4
requests==2.32.3
python-dotenv==1.0.1
prometheus-client==0.20.0
opencv-python==4.10.0.84
google-generativeai==0.8.3
onnxruntime==1.19.2
insightface==0.7.3
torch==2.4.1
torchvision==0.19.1
ultralytics==8.3.3
